backend_flask/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
gunicorn==23.0.0
SQLAlchemy==2.0.41
Flask-Migrate==4.1.0